+++ core/shared/src/main/scala/zio/ZIO.scala
@@ -474,38 +474,35 @@ sealed trait ZIO[-R, +E, +A] extends Serializable with ZIOPlatformSpecific[R, E,
   /**
    * Repeats this effect until its result satisfies the specified predicate.
    */
-  final def doUntil(f: A => Boolean): ZIO[R, E, A] =
-    repeat(Schedule.doUntil(f))
+  final def doUntil(f: A => Boolean): ZIO[R, E, A] = doUntilM(a => ZIO.succeed(f(a)))
 
   /**
    * Repeats this effect until its result is equal to the predicate.
    */
-  final def doUntilEquals[A1 >: A](a: => A1): ZIO[R, E, A1] =
-    repeat(Schedule.doUntilEquals(a))
+  final def doUntilEquals[A1 >: A](a: => A1): ZIO[R, E, A1] = doUntil(_ == a)
 
   /**
    * Repeats this effect until its result satisfies the specified effectful predicate.
    */
-  final def doUntilM(f: A => UIO[Boolean]): ZIO[R, E, A] =
-    repeat(Schedule.doUntilM(f))
+  final def doUntilM[R1 <: R](f: A => URIO[R1, Boolean]): ZIO[R1, E, A] =
+    self.flatMap(a => f(a).flatMap(b => if (b) ZIO.succeed(a) else self.doUntilM(f)))
 
   /**
    * Repeats this effect while its result satisfies the specified predicate.
    */
   final def doWhile(f: A => Boolean): ZIO[R, E, A] =
-    repeat(Schedule.doWhile(f))
+    doWhileM(a => ZIO.succeed(f(a)))
 
   /**
    * Repeats this effect for as long as the predicate is equal to its result.
    */
-  final def doWhileEquals[A1 >: A](a: => A1): ZIO[R, E, A1] =
-    repeat(Schedule.doWhileEquals(a))
+  final def doWhileEquals[A1 >: A](a: => A1): ZIO[R, E, A1] = doWhile(_ == a)
 
   /**
    * Repeats this effect while its result satisfies the specified effectful predicate.
    */
-  final def doWhileM(f: A => UIO[Boolean]): ZIO[R, E, A] =
-    repeat(Schedule.doWhileM(f))
+  final def doWhileM[R1 <: R](f: A => URIO[R1, Boolean]): ZIO[R1, E, A] =
+    self.flatMap(a => f(a).flatMap(b => if (!b) ZIO.succeed(a) else self.doUntilM(f)))
 
   /**
    * Returns an effect whose failure and success have been lifted into an
@@ -1474,7 +1471,7 @@ sealed trait ZIO[-R, +E, +A] extends Serializable with ZIOPlatformSpecific[R, E,
    * effect that executes `io`, and then if that succeeds, executes `io` an
    * additional time.
    */
-  final def repeat[R1 <: R, B](schedule: Schedule[R1, A, B]): ZIO[R1, E, B] =
+  final def repeat[R1 <: R, B](schedule: Schedule[R1, A, B]): ZIO[R1 with Clock, E, B] =
     repeatOrElse[R1, E, B](schedule, (e, _) => ZIO.fail(e))
 
   /**
@@ -1489,7 +1486,7 @@ sealed trait ZIO[-R, +E, +A] extends Serializable with ZIOPlatformSpecific[R, E,
   final def repeatOrElse[R1 <: R, E2, B](
     schedule: Schedule[R1, A, B],
     orElse: (E, Option[B]) => ZIO[R1, E2, B]
-  ): ZIO[R1, E2, B] =
+  ): ZIO[R1 with Clock, E2, B] =
     repeatOrElseEither[R1, B, E2, B](schedule, orElse).map(_.merge)
 
   /**
@@ -1504,22 +1501,30 @@ sealed trait ZIO[-R, +E, +A] extends Serializable with ZIOPlatformSpecific[R, E,
   final def repeatOrElseEither[R1 <: R, B, E2, C](
     schedule: Schedule[R1, A, B],
     orElse: (E, Option[B]) => ZIO[R1, E2, C]
-  ): ZIO[R1, E2, Either[C, B]] = {
-    def loop(last: A, state: schedule.State): ZIO[R1, E2, Either[C, B]] =
-      schedule
-        .update(last, state)
-        .foldM(
-          _ => ZIO.succeedRight(schedule.extract(last, state)),
-          s =>
-            self.foldM(
-              e => orElse(e, Some(schedule.extract(last, state))).map(Left(_)),
-              a => loop(a, s)
+  ): ZIO[R1 with Clock, E2, Either[C, B]] = {
+    import Schedule.Interval
+    import Schedule.Decision._
+    import java.util.concurrent.TimeUnit
+
+    def loop(optionB: Option[B], step: Schedule.StepFunction[R1, A, B]): ZIO[R1 with Clock, E2, Either[C, B]] =
+      self.foldM(
+        orElse(_, optionB).map(Left(_)),
+        a =>
+          clock
+            .currentTime(TimeUnit.MILLISECONDS)
+            .map(java.time.Instant.ofEpochMilli(_))
+            .flatMap(now =>
+              step(now, a).flatMap {
+                case Done(b) => ZIO.succeed(Right(b))
+                case Continue(b, interval, next) =>
+                  val duration = Interval(now, interval.start).size
+
+                  clock.sleep(duration) *> loop(Some(b), next)
+              }
             )
-        )
-    self.foldM(
-      orElse(_, None).map(Left(_)),
-      a => schedule.initial.flatMap(loop(a, _))
-    )
+      )
+
+    loop(None, schedule.step)
   }
 
   /**
@@ -1528,7 +1533,7 @@ sealed trait ZIO[-R, +E, +A] extends Serializable with ZIOPlatformSpecific[R, E,
    * `once` or `recurs` for example), so that that `io.retry(Schedule.once)` means
    * "execute `io` and in case of failure, try again once".
    */
-  final def retry[R1 <: R, S](policy: Schedule[R1, E, S])(implicit ev: CanFail[E]): ZIO[R1, E, A] =
+  final def retry[R1 <: R, S](policy: Schedule[R1, E, S])(implicit ev: CanFail[E]): ZIO[R1 with Clock, E, A] =
     retryOrElse(policy, (e: E, _: S) => ZIO.fail(e))
 
   /**
@@ -1539,67 +1544,77 @@ sealed trait ZIO[-R, +E, +A] extends Serializable with ZIOPlatformSpecific[R, E,
   final def retryOrElse[R1 <: R, A1 >: A, S, E1](
     policy: Schedule[R1, E, S],
     orElse: (E, S) => ZIO[R1, E1, A1]
-  )(implicit ev: CanFail[E]): ZIO[R1, E1, A1] =
+  )(implicit ev: CanFail[E]): ZIO[R1 with Clock, E1, A1] =
     retryOrElseEither(policy, orElse).map(_.merge)
 
   /**
-   * Retries with the specified schedule, until it fails, and then both the
-   * value produced by the schedule together with the last error are passed to
-   * the recovery function.
+   * Returns an effect that retries this effect with the specified schedule when it fails, until
+   * the schedule is done, then both the value produced by the schedule together with the last
+   * error are passed to the specified recovery function.
    */
-  final def retryOrElseEither[R1 <: R, S, E1, B](
-    policy: Schedule[R1, E, S],
-    orElse: (E, S) => ZIO[R1, E1, B]
-  )(implicit ev: CanFail[E]): ZIO[R1, E1, Either[B, A]] = {
-    def loop(state: policy.State): ZIO[R1, E1, Either[B, A]] =
-      self.foldM(
-        err =>
-          policy
-            .update(err, state)
-            .foldM(
-              _ => orElse(err, policy.extract(err, state)).map(Left(_)),
-              loop
-            ),
-        ZIO.succeedRight
-      )
+  final def retryOrElseEither[R1 <: R, Out, E1, B](
+    schedule: Schedule[R1, E, Out],
+    orElse: (E, Out) => ZIO[R1, E1, B]
+  )(implicit ev: CanFail[E]): ZIO[R1 with Clock, E1, Either[B, A]] = {
+    import Schedule.Interval
+    import Schedule.Decision._
+    import java.util.concurrent.TimeUnit
 
-    policy.initial.flatMap(loop)
+    def loop(step: Schedule.StepFunction[R1, E, Out]): ZIO[R1 with Clock, E1, Either[B, A]] =
+      self
+        .map(Right(_))
+        .catchAll(e =>
+          clock
+            .currentTime(TimeUnit.MILLISECONDS)
+            .map(java.time.Instant.ofEpochMilli(_))
+            .flatMap(now =>
+              step(now, e).flatMap {
+                case Done(out) => orElse(e, out).map(Left(_))
+                case Continue(_, interval, next) =>
+                  val duration = Interval(now, interval.start).size
+
+                  clock.sleep(duration) *> loop(next)
+              }
+            )
+        )
+
+    loop(schedule.step)
   }
 
   /**
    * Retries this effect until its error satisfies the specified predicate.
    */
-  final def retryUntil(f: E => Boolean)(implicit ev: CanFail[E]): ZIO[R, E, A] =
+  final def retryUntil(f: E => Boolean)(implicit ev: CanFail[E]): ZIO[R with Clock, E, A] =
     retry(Schedule.doUntil(f))
 
   /**
    * Retries this effect until its error equals the predicate.
    */
-  final def retryUntilEquals[E1 >: E](e: => E1)(implicit ev: CanFail[E1]): ZIO[R, E1, A] =
+  final def retryUntilEquals[E1 >: E](e: => E1)(implicit ev: CanFail[E1]): ZIO[R with Clock, E1, A] =
     retry(Schedule.doUntilEquals(e))
 
   /**
    * Retries this effect until its error satisfies the specified effectful predicate.
    */
-  final def retryUntilM(f: E => UIO[Boolean])(implicit ev: CanFail[E]): ZIO[R, E, A] =
+  final def retryUntilM[R1 <: R](f: E => URIO[R1, Boolean])(implicit ev: CanFail[E]): ZIO[R1 with Clock, E, A] =
     retry(Schedule.doUntilM(f))
 
   /**
    * Retries this effect while its error satisfies the specified predicate.
    */
-  final def retryWhile(f: E => Boolean)(implicit ev: CanFail[E]): ZIO[R, E, A] =
+  final def retryWhile(f: E => Boolean)(implicit ev: CanFail[E]): ZIO[R with Clock, E, A] =
     retry(Schedule.doWhile(f))
 
   /**
    * Repeats this effect for as long as the error equals the predicate.
    */
-  final def retryWhileEquals[E1 >: E](e: => E1)(implicit ev: CanFail[E1]): ZIO[R, E1, A] =
+  final def retryWhileEquals[E1 >: E](e: => E1)(implicit ev: CanFail[E1]): ZIO[R with Clock, E1, A] =
     retry(Schedule.doWhileEquals(e))
 
   /**
    * Retries this effect while its error satisfies the specified effectful predicate.
    */
-  final def retryWhileM(f: E => UIO[Boolean])(implicit ev: CanFail[E]): ZIO[R, E, A] =
+  final def retryWhileM[R1 <: R](f: E => URIO[R1, Boolean])(implicit ev: CanFail[E]): ZIO[R1 with Clock, E, A] =
     retry(Schedule.doWhileM(f))
 
   /**
